Separate components into different files
